Refactor auth
Do not clone the incomming request, but only the cookies

This is less hacky and works for os3
